remove handler autoconfiguration (#238)
