+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Removed auto configuration of default handlers.
+
 ## [0.10.4] 2020-03-20
 ### Added
 - Instrumented Resolvers with timers in line with queries and mutations
